requirements.txt
Django==5.0.7
python-decouple==3.8
psycopg2-binary==2.9.9
django-redis==6.0.0
PyJWT==2.10.1
drf-yasg==1.20.0